rearrange files to support examples
