+++ src/retomic/core.ts
@@ -1,0 +1,155 @@
+import {
+  useContext,
+  useEffect,
+  useMemo,
+  useState
+} from 'react'
+import { RootContext } from './constants'
+import { getState, setState } from './db'
+import { useLifeCycle } from './lifecycle'
+import { mutable } from './mutable'
+import type { AtomRef, WatcherFn } from './types'
+
+function defaultTo<T>(defaultValue: T, value: T) {
+  return value === undefined ? defaultValue : value
+}
+
+function $$resetAtom<T>(_: T, defaultState: T) {
+  return defaultState
+}
+
+function checkDuplicateAtomKey(key: AtomRef<any>['key']) {
+  const isDuplicateKey = mutable.atomRefsByKey.has(key)
+
+  if (isDuplicateKey) {
+    const duplicateKeyPrefix =
+      process.env.NODE_ENV === 'development'
+        ? '/@atomDuplicate'
+        : ''
+    const newKey = `${key}${duplicateKeyPrefix}/${mutable.duplicaKeyCount}`
+
+    mutable.duplicaKeyCount += 1
+    console.warn(
+      `Warning: duplicate atomRef key \`${key}\` detected. As a safety precaution a new key, \`${newKey}\`, was automatically generated.`
+    )
+
+    return newKey
+  }
+
+  return key
+}
+
+export type { AtomRef } from './types'
+export { useIsNew } from './utils'
+export { AtomDevTools } from './AtomDevTools'
+export { AtomRoot } from './AtomRoot'
+
+export function atomRef<T>({
+  key,
+  defaultState
+}: {
+  key: AtomRef<T>['key']
+  defaultState: AtomRef<T>['defaultState']
+}): Readonly<AtomRef<T>> {
+  const actualKey = checkDuplicateAtomKey(key)
+  const ref = {
+    key: actualKey,
+    defaultState
+  }
+
+  mutable.atomRefsByKey.set(actualKey, ref)
+
+  return ref
+}
+
+export function useReadAtom<T, SelectorValue = T>(
+  atomRef: AtomRef<T>,
+  selector: (state: T) => SelectorValue
+) {
+  const { key, defaultState } = atomRef
+  const rootDb = useContext(RootContext)
+  const initialStateSlice = getState(rootDb)[key]
+  const [hookState, setHookState] = useState(
+    selector(defaultTo(defaultState, initialStateSlice))
+  )
+
+  useEffect(() => {
+    const watcherFn: WatcherFn = ({ newState }) => {
+      const stateSlice = newState[key]
+      const nextValue = selector(
+        defaultTo(defaultState, stateSlice)
+      )
+      const hasChanged = hookState !== nextValue
+
+      if (!hasChanged) {
+        return
+      }
+
+      setHookState(nextValue)
+    }
+
+    return rootDb.subscriptions.on(key, watcherFn)
+  }, [
+    rootDb,
+    key,
+    hookState,
+    selector,
+    defaultState,
+    atomRef
+  ])
+  useLifeCycle(rootDb, atomRef)
+
+  return hookState
+}
+
+export function useSendAtom<T>(atomRef: AtomRef<T>) {
+  const { key, defaultState } = atomRef
+  const rootDb = useContext(RootContext)
+
+  useLifeCycle(rootDb, atomRef)
+  return useMemo(
+    () =>
+      <Payload>(
+        mutationFn: (oldState: T, payload: Payload) => T,
+        payload: Payload
+      ) => {
+        if (
+          process.env.NODE_ENV === 'development' &&
+          !mutationFn.name
+        ) {
+          console.error(
+            'Warning: This mutation function should be named -',
+            mutationFn
+          )
+        }
+
+        const rootState = getState(rootDb)
+        const stateSlice = defaultTo(
+          defaultState,
+          rootState[key]
+        )
+        const nextState = {
+          ...rootState,
+          [key]: mutationFn(stateSlice, payload)
+        }
+
+        return setState(
+          rootDb,
+          nextState,
+          atomRef,
+          mutationFn,
+          payload
+        )
+      },
+    [defaultState, rootDb, key, atomRef]
+  )
+}
+
+export function useResetAtom<T>(atomRef: AtomRef<T>) {
+  const mutate = useSendAtom(atomRef)
+
+  return useMemo(
+    () => () => mutate($$resetAtom, atomRef.defaultState),
+    [mutate, atomRef.defaultState]
+  )
+}
